analytics/requirements.txt
pathlib
psycopg2-binary
openpyxl